requirements.txt
playwright==1.40.0
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
